refactor move system commands to systems.py remove register_command function rework help command to make it less hacky
